requirements.txt
Django==4.1.5
djangorestframework==3.14.0
pytz==2022.1
requests==2.25.1
